Week3/Day4 - Create helpers.js file
