+++ views/partials/_header.ejs
@@ -6,16 +6,16 @@
     </button>
     <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
       <div class="navbar-nav">
+        <% if (userDB === null) { %> 
+        <a class="nav-item nav-link" href="/login">Login</a>
+        <a class="nav-item nav-link" href="/register">Register</a>
+        <% } else { %>
+          <a class="nav-item nav-link" href="/urls">My URLs</a>
+          <a class="nav-item nav-link" href="/urls/new">Create New URL</a>
          
-          <% if (userDB === null) { %> 
-          <a class="nav-item nav-link" href="/login">Login</a>
-          <a class="nav-item nav-link" href="/register">Register</a>
-          <% } else { %>
-            <a class="nav-item nav-link" href="/urls">My URLs</a>
-            <a class="nav-item nav-link" href="/urls/new">Create New URL</a>
-          <a class="nav-item nav-link">Logged in As: <%= userDB['email'] %></a><form method="POST" action="/logout"><button class="btn btn-outline-dark">Logout</button></form>
-        <% } %>
       </div>
     </div>
+    <a class="nav-item nav-link">Logged in As: <%= userDB['email'] %> </a><form method="POST" action="/logout"><button class="btn btn-outline-dark">Logout</button></form>
+  <% } %>
   </nav>
 </header>
