+++ views/urls_login.ejs
@@ -16,7 +16,7 @@
 
     <main style="margin: 1em;">
       <h3>Login</h3>
-      <form class="form-inline" action="/login" method="POST">
+      <form  action="/login" method="POST">
         <div class="form-group mb-2">
           <table aling='center'>
             <tr>
